SB-1 Added the interface and moved package
